wip: test otter.nvim performance, profile
