+++ lua/config/global.lua
@@ -1,6 +1,9 @@
 vim.g.markdown_fenced_languages = { 'html', 'python', 'bash=sh', 'R=r' }
 vim.opt.termguicolors = true
 
+vim.opt.shada = nil
+vim.opt.swapfile = false
+
 -- more opinionated
 vim.opt.number = true -- show linenumbers
 vim.opt.timeoutlen = 400 -- until which-key pops up
