+++ lua/plugins/quarto.lua
@@ -1,13 +1,397 @@
 return {
+    { 'nvim-treesitter/nvim-treesitter',
+      tag = nil,
+      branch = 'master',
+      run = ':TSUpdate',
+      config = function()
+        require 'nvim-treesitter.configs'.setup {
+          ensure_installed = { 'r', 'python', 'markdown', 'markdown_inline', 'julia', 'bash', 'yaml', 'lua', 'vim',
+            'query', 'vimdoc', 'latex' },
+          highlight = {
+            enable = true,
+            -- additional_vim_regex_highlighting = { 'markdown' },
+            additional_vim_regex_highlighting = false,
+            -- note: the vim regex based highlighting from
+            -- quarto-vim / vim-pandoc sets the wrong comment character
+            -- for some sections where there is `$` math.
+          },
+          indent = {
+            enable = true,
+          },
+          incremental_selection = {
+            enable = true,
+            keymaps = {
+              init_selection = "gnn",
+              node_incremental = "grn",
+              scope_incremental = "grc",
+              node_decremental = "grm",
+            },
+          },
+          textobjects = {
+            select = {
+              enable = true,
+              lookahead = true, -- Automatically jump forward to textobj, similar to targets.vim
+              keymaps = {
+                -- You can use the capture groups defined in textobjects.scm
+                ['af'] = '@function.outer',
+                ['if'] = '@function.inner',
+                ['ac'] = '@class.outer',
+                ['ic'] = '@class.inner',
+                ['ao'] = '@codechunk.outer',
+                ['io'] = '@codechunk.inner',
+              },
+            },
+            move = {
+              enable = true,
+              set_jumps = true, -- whether to set jumps in the jumplist
+              goto_next_start = {
+                [']m'] = '@function.outer',
+                [']c'] = '@codechunk.inner',
+                [']]'] = '@class.outer',
+              },
+              goto_next_end = {
+                [']M'] = '@function.outer',
+                [']['] = '@class.outer',
+              },
+              goto_previous_start = {
+                ['[m'] = '@function.outer',
+                ['[c'] = '@codechunk.inner',
+                -- ['[['] = '@class.outer',
+              },
+              goto_previous_end = {
+                ['[M'] = '@function.outer',
+                ['[]'] = '@class.outer',
+              },
+            },
+          },
+        }
+      end
+    },
+    -- { 'nvim-treesitter/nvim-treesitter-textobjects' },
+    -- { 'nvim-treesitter/playground',
+    --   tag = nil,
+    --   branch = 'master',
+    -- },
+  { 'neovim/nvim-lspconfig',
+    tag = nil,
+    version = nil,
+    event = "BufReadPre",
+    dependencies = {
+      -- { "folke/neoconf.nvim", cmd = "Neoconf", config = true },
+      -- { "folke/neodev.nvim", config = true },
+      { "williamboman/mason.nvim", config = true },
+      { "williamboman/mason-lspconfig.nvim",
+        tag = nil,
+        version = nil,
+        opts = {
+          automatic_installation = true,
+        }
+      },
+      { "hrsh7th/cmp-nvim-lsp" },
+    },
+    config = function()
+      local lspconfig = require('lspconfig')
+      local cmp_nvim_lsp = require('cmp_nvim_lsp')
+      local util = require("lspconfig.util")
+
+      local on_attach = function(client, bufnr)
+        local function buf_set_keymap(...) vim.api.nvim_buf_set_keymap(bufnr, ...) end
+
+        local function buf_set_option(...) vim.api.nvim_buf_set_option(bufnr, ...) end
+
+        buf_set_option('omnifunc', 'v:lua.vim.lsp.omnifunc')
+        local opts = { noremap = true, silent = true }
+
+        buf_set_keymap('n', 'gD', '<cmd>Telescope lsp_type_definitions<CR>', opts)
+        buf_set_keymap('n', 'gd', '<cmd>Telescope lsp_definitions<CR>', opts)
+        buf_set_keymap('n', 'K', '<cmd>lua vim.lsp.buf.hover()<CR>', opts)
+        buf_set_keymap('n', 'gh', '<cmd>lua vim.lsp.buf.signature_help()<CR>', opts)
+        buf_set_keymap('n', 'gi', '<cmd>Telescope lsp_implementations<CR>', opts)
+        buf_set_keymap('n', 'gr', '<cmd>Telescope lsp_references<CR>', opts)
+        buf_set_keymap('n', '[d', '<cmd>lua vim.diagnostic.goto_prev()<CR>', opts)
+        buf_set_keymap('n', ']d', '<cmd>lua vim.diagnostic.goto_next()<CR>', opts)
+        buf_set_keymap('n', '<leader>ll', '<cmd>lua vim.lsp.codelens.run()<cr>', opts)
+        client.server_capabilities.document_formatting = true
+      end
+
+
+      local on_attach2 = function(client, bufnr)
+        local function buf_set_keymap(...) vim.api.nvim_buf_set_keymap(bufnr, ...) end
+
+        local function buf_set_option(...) vim.api.nvim_buf_set_option(bufnr, ...) end
+
+        buf_set_option('omnifunc', 'v:lua.vim.lsp.omnifunc')
+        local opts = { noremap = true, silent = true }
+
+        buf_set_keymap('n', 'gD', '<cmd>Telescope lsp_type_definitions<CR>', opts)
+        buf_set_keymap('n', 'gh', '<cmd>lua vim.lsp.buf.signature_help()<CR>', opts)
+        buf_set_keymap('n', 'gi', '<cmd>Telescope lsp_implementations<CR>', opts)
+        buf_set_keymap('n', 'gr', '<cmd>Telescope lsp_references<CR>', opts)
+        buf_set_keymap('n', '[d', '<cmd>lua vim.diagnostic.goto_prev()<CR>', opts)
+        buf_set_keymap('n', ']d', '<cmd>lua vim.diagnostic.goto_next()<CR>', opts)
+        buf_set_keymap('n', '<leader>ll', '<cmd>lua vim.lsp.codelens.run()<cr>', opts)
+        client.server_capabilities.document_formatting = true
+      end
+
+      local lsp_flags = {
+        allow_incremental_sync = true,
+        debounce_text_changes = 150,
+      }
+
+      vim.lsp.handlers['textDocument/publishDiagnostics'] = vim.lsp.with(vim.lsp.diagnostic.on_publish_diagnostics, {
+        virtual_text = true,
+        signs = true,
+        underline = true,
+        update_in_insert = false,
+      })
+      vim.lsp.handlers["textDocument/hover"] = vim.lsp.with(vim.lsp.handlers.hover,
+        { border = require 'misc.style'.border })
+      vim.lsp.handlers["textDocument/signatureHelp"] = vim.lsp.with(vim.lsp.handlers.signature_help,
+        { border = require 'misc.style'.border })
+
+      local capabilities = vim.lsp.protocol.make_client_capabilities()
+      capabilities = cmp_nvim_lsp.default_capabilities(capabilities)
+      capabilities.textDocument.completion.completionItem.snippetSupport = true
+
+      -- also needs:
+      -- $home/.config/marksman/config.toml :
+      -- [core]
+      -- markdown.file_extensions = ["md", "markdown", "qmd"]
+      lspconfig.marksman.setup {
+        on_attach = on_attach2,
+        capabilities = capabilities,
+        filetypes = { 'markdown', 'quarto' },
+        root_dir = util.root_pattern(".git", ".marksman.toml", "_quarto.yml"),
+      }
+
+      lspconfig.r_language_server.setup {
+        on_attach = on_attach,
+        capabilities = capabilities,
+        flags = lsp_flags
+      }
+
+      lspconfig.emmet_ls.setup {
+        on_attach = on_attach,
+        capabilities = capabilities,
+        flags = lsp_flags
+      }
+
+      lspconfig.cssls.setup {
+        on_attach = on_attach,
+        capabilities = capabilities,
+        flags = lsp_flags
+      }
+
+      local function strsplit(s, delimiter)
+        local result = {}
+        for match in (s .. delimiter):gmatch("(.-)" .. delimiter) do
+          table.insert(result, match)
+        end
+        return result
+      end
+
+      local function get_quarto_resource_path()
+        local f = assert(io.popen('quarto --paths', 'r'))
+        local s = assert(f:read('*a'))
+        f:close()
+        return strsplit(s, '\n')[2]
+      end
+
+      local lua_library_files = vim.api.nvim_get_runtime_file("", true)
+      local resource_path = get_quarto_resource_path()
+      table.insert(lua_library_files, resource_path .. '/lua-types')
+      local lua_plugin_paths = {}
+      table.insert(lua_plugin_paths, resource_path .. '/lua-plugin/plugin.lua')
+
+      lspconfig.lua_ls.setup {
+        on_attach = on_attach,
+        capabilities = capabilities,
+        flags = lsp_flags,
+        settings = {
+          Lua = {
+            completion = {
+              callSnippet = "Replace"
+            },
+            runtime = {
+              version = 'LuaJIT',
+              plugin = lua_plugin_paths,
+            },
+            diagnostics = {
+              globals = { 'vim', 'quarto', 'pandoc', 'io', 'string', 'print', 'require', 'table' },
+              disable = { 'trailing-space' },
+            },
+            workspace = {
+              library = lua_library_files,
+              checkThirdParty = false,
+            },
+            telemetry = {
+              enable = false,
+            },
+          },
+        },
+      }
+
+      lspconfig.pyright.setup {
+        on_attach = on_attach,
+        capabilities = capabilities,
+        flags = lsp_flags,
+        root_dir = function(fname)
+          return util.root_pattern(".git", "setup.py", "setup.cfg", "pyproject.toml", "requirements.txt")(fname) or
+              util.path.dirname(fname)
+        end
+      }
+
+      lspconfig.julials.setup {
+        on_attach = on_attach,
+        capabilities = capabilities,
+        flags = lsp_flags,
+      }
+
+      lspconfig.bashls.setup {
+        on_attach = on_attach,
+        capabilities = capabilities,
+        flags = lsp_flags,
+        filetypes = { 'sh', 'bash' }
+      }
+    end
+  },
+
+
+  -- completion
+  { 'hrsh7th/nvim-cmp',
+    dependencies = {
+      { 'hrsh7th/cmp-nvim-lsp' },
+      { 'hrsh7th/cmp-nvim-lsp-signature-help' },
+      { 'hrsh7th/cmp-buffer' },
+      { 'hrsh7th/cmp-path' },
+      { 'hrsh7th/cmp-calc' },
+      { 'hrsh7th/cmp-emoji' },
+      { 'saadparwaiz1/cmp_luasnip' },
+      { 'f3fora/cmp-spell' },
+      { 'ray-x/cmp-treesitter' },
+      { 'kdheepak/cmp-latex-symbols' },
+      -- { 'jc-doyle/cmp-pandoc-references' },
+      { 'jmbuhr/cmp-pandoc-references' },
+      { 'L3MON4D3/LuaSnip' },
+      { 'rafamadriz/friendly-snippets' },
+      { 'onsails/lspkind-nvim' },
+    },
+    config = function()
+      local cmp = require 'cmp'
+      local luasnip = require 'luasnip'
+      local lspkind = require "lspkind"
+      lspkind.init()
+
+
+      local has_words_before = function()
+        local line, col = unpack(vim.api.nvim_win_get_cursor(0))
+        return col ~= 0 and vim.api.nvim_buf_get_lines(0, line - 1, line, true)[1]:sub(col, col):match("%s") == nil
+      end
+
+      cmp.setup({
+        snippet = {
+          expand = function(args)
+            luasnip.lsp_expand(args.body)
+          end,
+        },
+        mapping = {
+          ['<C-f>'] = cmp.mapping.scroll_docs(-4),
+          ['<C-d>'] = cmp.mapping.scroll_docs(4),
+          ['<C-n>'] = cmp.mapping(function(fallback)
+            if luasnip.expand_or_jumpable() then
+              luasnip.expand_or_jump()
+              fallback()
+            end
+          end, { "i", "s" }),
+          ['<C-p>'] = cmp.mapping(function(fallback)
+            if luasnip.jumpable(-1) then
+              luasnip.jump(-1)
+            else
+              fallback()
+            end
+          end, { "i", "s" }),
+          ['<c-a>'] = cmp.mapping.complete(),
+          ['<C-e>'] = cmp.mapping.abort(),
+          ['<CR>'] = cmp.mapping.confirm({
+            select = true,
+          }),
+          ["<Tab>"] = cmp.mapping(function(fallback)
+            if cmp.visible() then
+              cmp.select_next_item()
+            elseif has_words_before() then
+              cmp.complete()
+            else
+              fallback()
+            end
+          end, { "i", "s" }),
+          ["<S-Tab>"] = cmp.mapping(function(fallback)
+            if cmp.visible() then
+              cmp.select_prev_item()
+            else
+              fallback()
+            end
+          end, { "i", "s" }),
+        },
+        autocomplete = false,
+        formatting = {
+          format = lspkind.cmp_format {
+            with_text = true,
+            menu = {
+              otter = "[🦦]",
+              luasnip = "[snip]",
+              nvim_lsp = "[LSP]",
+              buffer = "[buf]",
+              path = "[path]",
+              spell = "[spell]",
+              pandoc_references = "[ref]",
+              tags = "[tag]",
+              treesitter = "[TS]",
+              calc = "[calc]",
+              latex_symbols = "[tex]",
+              emoji = "[emoji]",
+            },
+          },
+        },
+        sources = {
+          { name = 'otter' }, -- quarto
+          { name = 'path' },
+          { name = 'nvim_lsp' },
+          { name = 'nvim_lsp_signature_help' },
+          { name = 'luasnip', keyword_length = 3, max_item_count = 3 },
+          { name = 'pandoc_references' },
+          { name = 'buffer', keyword_length = 5, max_item_count = 3 },
+          { name = 'spell' },
+          { name = 'treesitter', keyword_length = 5, max_item_count = 3 },
+          { name = 'calc' },
+          { name = 'latex_symbols' },
+          { name = 'emoji' },
+        },
+        view = {
+          entries = "native",
+        },
+        window = {
+          documentation = {
+            border = require 'misc.style'.border,
+          },
+        },
+      })
+      -- for friendly snippets
+      require("luasnip.loaders.from_vscode").lazy_load()
+      -- for custom snippets
+      require("luasnip.loaders.from_vscode").lazy_load({ paths = { "~/.config/nvim/snips" } })
+
+    end
+  },
+
+
   { 'quarto-dev/quarto-nvim',
-    dev = false,
+    dev = true,
     dependencies = {
       { 'hrsh7th/nvim-cmp' },
-      { 'neovim/nvim-lspconfig' },
-      { 'nvim-treesitter/nvim-treesitter' },
       { 'jmbuhr/otter.nvim',
-        tag = nil,
-        branch = 'nightly-treesitter',
+        dev = true,
+        -- tag = nil,
+        -- branch = 'nightly-treesitter',
         config = function()
           require 'otter.config'.setup {
             lsp = {
@@ -18,15 +402,12 @@ return {
           }
         end,
       },
-      { 'quarto-dev/quarto-vim',
-        ft = 'quarto',
-        dependencies = { 'vim-pandoc/vim-pandoc-syntax' },
-        -- note: needs additional vim highlighting enabled
-        -- for markdown in treesitter.lua
-      },
-    },
-    config = function()
-
+      -- { 'quarto-dev/quarto-vim',
+      --   ft = 'quarto',
+      --   dependencies = { 'vim-pandoc/vim-pandoc-syntax' },
+      --   -- note: needs additional vim highlighting enabled
+      --   -- for markdown in treesitter.lua
+      --   config = function()
       -- conceal can be tricky because both
       -- the treesitter highlighting and the
       -- regex vim syntax files can define conceals
@@ -42,29 +423,33 @@ return {
       -- 		custom replacement character defined (see
       -- 		|:syn-cchar|).
       -- 3		Concealed text is completely hidden.
-      vim.opt.conceallevel = 1
-
-      -- disable conceal in markdown/quarto
-      vim.g['pandoc#syntax#conceal#use'] = false
-
-      -- embeds are already handled by treesitter injectons
-      vim.g['pandoc#syntax#codeblocks#embeds#use'] = false
-
-      vim.g['pandoc#syntax#conceal#blacklist'] = { 'codeblock_delim', 'codeblock_start' }
-
-      -- but allow some types of conceal in math reagions:
-      -- a=accents/ligatures d=delimiters m=math symbols
-      -- g=Greek  s=superscripts/subscripts
-      vim.g['tex_conceal'] = 'gm'
-
+    --
+      -- vim.opt.conceallevel = 1
+      --
+      -- -- disable conceal in markdown/quarto
+      -- vim.g['pandoc#syntax#conceal#use'] = false
+      --
+      -- -- embeds are already handled by treesitter injectons
+      -- vim.g['pandoc#syntax#codeblocks#embeds#use'] = false
+      --
+      -- vim.g['pandoc#syntax#conceal#blacklist'] = { 'codeblock_delim', 'codeblock_start' }
+      --
+      -- -- but allow some types of conceal in math reagions:
+      -- -- a=accents/ligatures d=delimiters m=math symbols
+      -- -- g=Greek  s=superscripts/subscripts
+      -- vim.g['tex_conceal'] = 'gm'
+      -- --   end
+      -- },
+    },
+    config = function()
       require 'quarto'.setup {
         lspFeatures = {
           enabled = true,
           languages = { 'r', 'python', 'julia', 'bash' },
           chunks = 'curly', -- 'curly' or 'all'
           diagnostics = {
-            enabled = true,
-            triggers = { "BufWrite" }
+            enabled = false,
+            triggers = { "BufWritePost" }
           },
           completion = {
             enabled = true
@@ -79,52 +464,52 @@ return {
   { 'jpalardy/vim-slime',
     init = function()
 
-      Quarto_is_in_python_chunk = function()
-        vim.b.quarto_is_python_chunk = false
-        local ts = vim.treesitter
-        local language_tree = ts.get_parser(0, 'markdown')
-        local syntax_tree = language_tree:parse()
-        local root = syntax_tree[1]:root()
-
-        -- create capture
-        local query = vim.treesitter.query.parse('markdown', require 'otter.tools.queries'['markdown'])
-
-        local row, col = unpack(vim.api.nvim_win_get_cursor(0))
-        row = row - 1
-        col = col
-
-        -- get text ranges
-        for pattern, match, metadata in query:iter_matches(root, 0) do
-          -- each match has two nodes, the language and the code
-          -- the language node is the first one
-          local found = false -- reset found for the next match
-          for id, node in pairs(match) do
-            local name = query.captures[id]
-            local ok, text = pcall(vim.treesitter.get_node_text, node, 0)
-            if not ok then return false end
-            if name == 'lang' and text == 'python' then
-              -- we found a match where the language node matches
-              -- the otter language
-              found = true
-            end
-            -- the corresponding code is in the current range
-            if found and name == 'code' and ts.is_in_node_range(node, row, col) then
-              vim.b.quarto_is_python_chunk = true
-            end
-          end
-        end
-      end
-
-      vim.cmd [[
-      function SlimeOverride_EscapeText_quarto(text)
-        call v:lua.Quarto_is_in_python_chunk() 
-        if exists('g:slime_python_ipython') && len(split(a:text,"\n")) > 1 && b:quarto_is_python_chunk
-          return ["%cpaste -q", "\n", g:slime_dispatch_ipython_pause, a:text, "--", "\n"]
-        else
-          return a:text
-        end
-      endfunction
-      ]]
+      -- Quarto_is_in_python_chunk = function()
+      --   vim.b.quarto_is_python_chunk = false
+      --   local ts = vim.treesitter
+      --   local language_tree = ts.get_parser(0, 'markdown')
+      --   local syntax_tree = language_tree:parse()
+      --   local root = syntax_tree[1]:root()
+      --
+      --   -- create capture
+      --   local query = vim.treesitter.query.parse('markdown', require 'otter.tools.queries'['markdown'])
+      --
+      --   local row, col = unpack(vim.api.nvim_win_get_cursor(0))
+      --   row = row - 1
+      --   col = col
+      --
+      --   -- get text ranges
+      --   for pattern, match, metadata in query:iter_matches(root, 0) do
+      --     -- each match has two nodes, the language and the code
+      --     -- the language node is the first one
+      --     local found = false -- reset found for the next match
+      --     for id, node in pairs(match) do
+      --       local name = query.captures[id]
+      --       local ok, text = pcall(vim.treesitter.get_node_text, node, 0)
+      --       if not ok then return false end
+      --       if name == 'lang' and text == 'python' then
+      --         -- we found a match where the language node matches
+      --         -- the otter language
+      --         found = true
+      --       end
+      --       -- the corresponding code is in the current range
+      --       if found and name == 'code' and ts.is_in_node_range(node, row, col) then
+      --         vim.b.quarto_is_python_chunk = true
+      --       end
+      --     end
+      --   end
+      -- end
+      --
+      -- vim.cmd [[
+      -- function SlimeOverride_EscapeText_quarto(text)
+      --   call v:lua.Quarto_is_in_python_chunk() 
+      --   if exists('g:slime_python_ipython') && len(split(a:text,"\n")) > 1 && b:quarto_is_python_chunk
+      --     return ["%cpaste -q", "\n", g:slime_dispatch_ipython_pause, a:text, "--", "\n"]
+      --   else
+      --     return a:text
+      --   end
+      -- endfunction
+      -- ]]
 
       vim.b.slime_cell_delimiter = "#%%"
       -- -- slime, tmux
@@ -141,12 +526,5 @@ return {
   -- :PasteImg,
   'ekickx/clipboard-image.nvim',
 
-  -- display images in the terminal!
-  { 'edluffy/hologram.nvim',
-    config = function()
-      -- require'hologram'.setup{
-      --   auto_display = true
-      -- }
-    end
-  },
+
 }
